충돌 판정 이슈 수정

+++ src/Math.cpp
@@ -45,7 +45,7 @@ Vector2 Vector2::normalized() const {
 }
 
 float Math::distance(const Vector2& a, const Vector2& b) {
-  return (float)std::sqrt(std::fabsf(a.x - b.x) * std::fabsf(a.x - b.x) +
+  return (float)std::sqrtf(std::fabsf(a.x - b.x) * std::fabsf(a.x - b.x) +
                           std::fabsf(a.y - b.y) * std::fabsf(a.y - b.y));
 }
